moving some files around and beginning to link components up to containers
